Moved callOncePerLoop in parallel/util/. Renamed stream synchronize method in TpAccumulator.
