+++ test/unit/util/CMakeLists.txt
@@ -1,6 +1,5 @@
 # util's unit tests
 
-dca_add_gtest(call_once_per_loop_test GTEST_MAIN LIBS parallel_stdthread)
 dca_add_gtest(ignore_test GTEST_MAIN)
 dca_add_gtest(integer_division_test GTEST_MAIN)
 dca_add_gtest(print_type_test GTEST_MAIN)
